% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHDGridlist.R
\name{getHDGridlist}
\alias{getHDGridlist}
\title{Reads LPJ-GUESS half-degree gridlist}
\usage{
getHDGridlist(type = "coords", SW.corner = FALSE)
}
\arguments{
\item{type}{A character specifying "coords" which gives the longitudes and latitides as coordinates,
"index" which gives the indices of the gridcells (compatible with the 'cf' input model)
or "both" which return both the coordinates indices}

\item{SW.corner}{A logical, if TRUE specify gridcells by the the south-west corner (useful for old LPJ-GUESS versions)}
}
\value{
A data.frame with all the standard global gridcells
}
\description{
Returns (as a data.frame) the standard 59,191 0.5 degree gridcell for a global LPJ-GUESS run as a data.frame.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
