% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUESS.Euro.PFTs.R
\docType{data}
\name{lpj.euro.PFTs}
\alias{lpj.euro.PFTs}
\title{dummy text}
\format{
A list of \code{PFT} objects that store meta-data for standard PFT for supported models
}
\usage{
lpj.euro.PFTs
}
\description{

}
\details{
LPJ-GUESS European PFTs
}
\keyword{datasets}
