% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGridlist.R
\name{subsetGridlist}
\alias{subsetGridlist}
\title{Select a subset of a gridlist}
\usage{
subsetGridlist(gridlist, area, file.name, header = TRUE, offset = c(0,
  0))
}
\arguments{
\item{gridlist}{A character string giving the location of the original gridlist which will be read by read.table and the first two columns will be
assumed to be Lon then Lat, _or_ a data.frame wth columns Lon and Lat.}

\item{area}{A raster::Extent or a SpatialExtent object specifying the geograpical sub-domain required}

\item{file.name}{A character string specifying a path to write the new gridlist (can be ignored to write no file)}

\item{header}{Logical, whether or not the original file has a header}

\item{offset}{A two-member numeric vector specifying the longitude-latitude offset from the given coordinates
to the gridcell centre, ie. should be c(0.25,0.25) for old LPJ-GUESS gridlists, and c(0.0,0.0) for most other more sensible scenarios}
}
\value{
The new gridlist as a data.frame (is also written to a file if file.name is provided)
}
\description{
Crop a gridlist to an smaller area.  Grridlist taken as a file in the form of a table of text (first and second columsn assume to be longitude and latitude repsective) or a data.frame with columns "Lon" and "Lat".
The smaller area is anything from which a raster::Extent can be derived.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
